+++ HeavenClient/Gameplay/MapleMap/MapMobs.cpp
@@ -129,12 +129,12 @@ namespace ms
 		Range<int16_t> horizontal = Range<int16_t>(moveobj.get_last_x(), moveobj.get_x());
 		Range<int16_t> vertical = Range<int16_t>(moveobj.get_last_y(), moveobj.get_y());
 
-		Rectangle<int16_t> player_rect = {
-				horizontal.smaller(),
-				horizontal.greater(),
-				vertical.smaller() - 50,
-				vertical.greater()
-		};
+			Rectangle<int16_t> player_rect = {
+					horizontal.smaller(),
+					horizontal.greater(),
+					static_cast<int16_t>(vertical.smaller() - 50),
+					vertical.greater()
+			};
 
 		auto iter = std::find_if(
 				mobs.begin(),
